refactor common context
